move 'pipeP' tests to separate file
