+++ types/ramda/ramda-tests.ts
@@ -126,31 +126,6 @@ class F2 {
     const b: string = truncate("0123456789ABC"); // => '0123456789…'
 };
 
-/* pipeP */
-() => {
-    interface User {
-        followers: string[];
-        name: string;
-    }
-
-    const db = {
-        getUserById(userName: string): Promise<User> {
-            return Promise.resolve({
-                name: 'Jon',
-                followers: [
-                    'Samwell',
-                    'Edd',
-                    'Grenn',
-                ],
-            });
-        },
-        getFollowers(user: User): Promise<string[]> {
-            return Promise.resolve(user.followers);
-        },
-    };
-    const followersForUser: (userName: string) => Promise<string[]> = R.pipeP(db.getUserById, db.getFollowers);
-};
-
 () => {
     interface User {
         followers: string[];
