+++ types/ramda/tsconfig.json
@@ -174,6 +174,7 @@
         "test/pickBy-tests.ts",
         "test/pipe-tests.ts",
         "test/pipeK-tests.ts",
+        "test/pipeP-tests.ts",
         "test/propEq-tests.ts"
     ]
 }
